+++ bld/wv/qnx/c/qnxfilio.c
@@ -90,7 +90,7 @@ sys_handle LocalOpen( const char *name, obj_attrs oattrs )
         SET_SYSHANDLE_NULL( sh );
         return( sh );
     }
-    fcntl( ret, F_SETFD, (int)FD_CLOEXEC );
+    fcntl( ret, F_SETFD, FD_CLOEXEC );
     LH2SYSH( sh, ret );
     return( sh );
 }
